requirements.txt
acres==0.5.0
aiofiles==24.1.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
audioop-lts==0.2.2
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
ci-info==0.3.0
click==8.3.0
cohere==5.20.0
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
cryptography==46.0.3
etelemetry==0.3.1
fastapi==0.121.0
fastavro==1.12.1
ffmpy==0.6.4
filelock==3.20.0
fsspec==2025.10.0
gradio==5.49.1
gradio_client==1.13.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==1.0.1
idna==3.11
influxdb-client==1.49.0
Jinja2==3.1.6
looseversion==1.3.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==2.3.4
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20250506
pillow==11.3.0
prov==2.1.1
puremagic==1.30
py4j==0.10.9.9
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
pydot==4.0.1
pydub==0.25.1
Pygments==2.19.2
PyMuPDF==1.26.5
pyparsing==3.2.5
pyspark==4.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-multipart==0.0.20
pytz==2025.2
pyxnat==1.6.3
PyYAML==6.0.3
rdflib==7.4.0
reactivex==4.0.4
requests==2.32.5
rich==14.2.0
ruff==0.14.3
safehttpx==0.1.7
scipy==1.16.3
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
tokenizers==0.22.1
tomlkit==0.13.3
tqdm==4.67.1
traits==7.0.2
typer==0.20.0
typer-slim==0.20.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
python-dotenv==1.2.1